+++ frontend/src/Landing.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import { Card, CardBody, CardHeader, CardTitle } from '@sveltestrap/sveltestrap'
   import { profile } from './api/profile.svelte'
-  import { theme } from './lib/theme.svelte'
+  import { theme } from './includes/theme.svelte'
 </script>
 
-<Card color="notifiarr" outline class="mb-2" {theme}>
+<Card color="notifiarr" outline class="mb-2" theme={$theme}>
   <CardHeader>
     <CardTitle>Welcome, {$profile?.username}!</CardTitle>
   </CardHeader>
