+++ frontend/src/includes/locale/de.json
