+++ frontend/src/includes/locale/es.json
