+++ frontend/src/includes/locale/fr.json
