+++ frontend/src/includes/locale/hu.json
