+++ frontend/src/includes/locale/index.svelte.ts
