+++ frontend/src/includes/locale/pl.json
