+++ frontend/src/includes/locale/sv.json
