+++ frontend/src/index.ts
@@ -1,6 +1,6 @@
 import { mount } from 'svelte'
-import '/src/app.css'
+import '/src/includes/app.css'
 import Index from '/src/Index.svelte'
-import '/src/lib/locale/index.svelte.ts'
+import '/src/includes/locale/index.svelte.ts'
 
 export default mount(Index, { target: document.getElementById('notifiarr')! })
